requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
streamlit==1.28.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
youtube-search-python==1.6.6
duckduckgo-search==3.9.10
openai==1.3.5
langchain==0.1.0
pydantic==2.5.0
